ex02

+++ ex01/Fixed.cpp
@@ -6,7 +6,7 @@
 /*   By: takawauc <takawauc@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/12/15 12:46:39 by takawauc          #+#    #+#             */
-/*   Updated: 2025/12/16 16:40:45 by takawauc         ###   ########.fr       */
+/*   Updated: 2025/12/16 16:47:45 by takawauc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ Fixed::Fixed(int integer) : _rawBits(integer << 8) {
 }
 
 Fixed::Fixed(double float_num)
-    : _rawBits(std::round(float_num * (1 << _frac_bits))) {
+    : _rawBits(std::roundf(float_num * (1 << _frac_bits))) {
   std::cout << "Float constructor called\n";
 }
